% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{applyTemplate}
\alias{applyTemplate}
\title{Applies a provided template string to a list of arguments.
Any argumements gives will be replaced in the template via placeholders wrapped in <<argument>>}
\usage{
applyTemplate(template, arguments = list())
}
\arguments{
\item{template}{A string with placeholders that can be replaced with the given arguments.}

\item{arguments}{Named list with values used for the template placeholders.}
}
\value{
A string based on the template with the diferent argumetns applied.
}
\description{
Applies a provided template string to a list of arguments.
Any argumements gives will be replaced in the template via placeholders wrapped in <<argument>>
}
