% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{createDirectories}
\alias{createDirectories}
\title{Generates the folder structure required for the pwa files.}
\usage{
createDirectories()
}
\description{
Generates the folder structure required for the pwa files.
}
\seealso{
[createServiceWorker()], [createIcon()], [createOfflinePage()],
   [createManifest()]

Other fileGen: 
\code{\link{createIcon}()},
\code{\link{createManifest}()},
\code{\link{createOfflinePage}()},
\code{\link{createServiceWorker}()}
}
\concept{fileGen}
