% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{createIcon}
\alias{createIcon}
\title{Creates the pwa icon file based on the given path.}
\usage{
createIcon(icon)
}
\arguments{
\item{icon}{Path to the icon used for PWA installations.}
}
\description{
Creates the pwa icon file based on the given path.
}
\seealso{
[createDirectories()], [createServiceWorker()],
   [createOfflinePage()], [createManifest()]

Other fileGen: 
\code{\link{createDirectories}()},
\code{\link{createManifest}()},
\code{\link{createOfflinePage}()},
\code{\link{createServiceWorker}()}
}
\concept{fileGen}
