% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{createManifest}
\alias{createManifest}
\title{Creates the manifest file.}
\usage{
createManifest(title, start_url, color)
}
\arguments{
\item{title}{title of the page when the pwa is running in offline mode.}

\item{start_url}{The ull url where the app is hosted.}

\item{color}{An argument that can be used in the
offline template to show a custom message.}
}
\description{
Creates the manifest file.
}
\seealso{
[createDirectories()], [createServiceWorker()], [createIcon()],
   [createOfflinePage()]

Other fileGen: 
\code{\link{createDirectories}()},
\code{\link{createIcon}()},
\code{\link{createOfflinePage}()},
\code{\link{createServiceWorker}()}
}
\concept{fileGen}
