% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{createOfflinePage}
\alias{createOfflinePage}
\title{Creates the offline landing page.}
\usage{
createOfflinePage(title, template, offline_message)
}
\arguments{
\item{title}{title of the page when the pwa is running in offline mode.}

\item{template}{HTML template to use. If null a default template is used.}

\item{offline_message}{An argument that can be used
in the offline template to show a custom message.}
}
\description{
Creates the offline landing page.
}
\seealso{
[createDirectories()], [createServiceWorker()], [createIcon()],
   [createManifest()]

Other fileGen: 
\code{\link{createDirectories}()},
\code{\link{createIcon}()},
\code{\link{createManifest}()},
\code{\link{createServiceWorker}()}
}
\concept{fileGen}
