% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{createServiceWorker}
\alias{createServiceWorker}
\title{Creates the service worker file based of the package template file.}
\usage{
createServiceWorker()
}
\description{
Creates the service worker file based of the package template file.
}
\seealso{
[createDirectories()], [createIcon()], [createOfflinePage()],
   [createManifest()]

Other fileGen: 
\code{\link{createDirectories}()},
\code{\link{createIcon}()},
\code{\link{createManifest}()},
\code{\link{createOfflinePage}()}
}
\concept{fileGen}
