% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{generateFiles}
\alias{generateFiles}
\title{Generates the required files for enabling the pwa.}
\usage{
generateFiles(
  title,
  domain,
  location,
  output,
  offline_template,
  offline_message,
  icon,
  color
)
}
\arguments{
\item{title}{The title of your Shiny app}

\item{domain}{the base URL where the app is hosted}

\item{location}{subdirectory where the app is hosted. Only required if the app is not on the root domain.}

\item{output}{Relative folder where to create the service worker file.
Usually corresponds to the folder used by shiny to serve static files,
this folder must exist and is usually the www folder of your shiny project.}

\item{offline_template}{Path to the offline template you want to use. If left NULL the default template is used.}

\item{offline_message}{When using the default offline page template, message to be displayed.}

\item{icon}{Icon Path to be used for the app. Size should be 512x512px. If left NULL a default icon is provided.}

\item{color}{Color of the app. Used to color the browser elements when the pwa is installed.}
}
\value{
A UI definition that can be passed to the [shinyUI] function.
}
\description{
Generates the required files for enabling the pwa.
}
