% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{getTemplate}
\alias{getTemplate}
\title{Get the full path for a default file template}
\usage{
getTemplate(file)
}
\arguments{
\item{file}{Name of the file including the file extension.}
}
\description{
Get the full path for a default file template
}
