% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{loadDependencies}
\alias{loadDependencies}
\title{Returns a set of tags required for registering the pwa.}
\usage{
loadDependencies(title, icon, color, location)
}
\arguments{
\item{title}{Title of the app}

\item{icon}{Name of the icon file to be user}

\item{color}{color used for the UI on the installed pwa}

\item{location}{subdirectory where the app is hosted. Only required if the app is not on the root domain.}
}
\value{
A UI definition that can be passed to the [shinyUI] function.
}
\description{
Returns a set of tags required for registering the pwa.
}
\seealso{
[generateFiles()]
}
\concept{dependencies}
