% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateDomain}
\alias{validateDomain}
\title{Clears and validates the provided domain.}
\usage{
validateDomain(domain)
}
\arguments{
\item{domain}{A string representing the domain of the application}
}
\value{
A validated version of the domain
}
\description{
Clears and validates the provided domain.
}
\seealso{
[validateLocation()], [validateIcon()]

Other validators: 
\code{\link{validateIcon}()},
\code{\link{validateLocation}()}
}
\concept{validators}
