% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateIcon}
\alias{validateIcon}
\title{Validates the provided icon. If the icon doesnt exist returns a default one.}
\usage{
validateIcon(icon)
}
\arguments{
\item{icon}{Path location for an icon relative to the project root}
}
\value{
A valid icon path
}
\description{
Validates the provided icon. If the icon doesnt exist returns a default one.
}
\seealso{
[validateDomain()], [validateLocation()]

Other validators: 
\code{\link{validateDomain}()},
\code{\link{validateLocation}()}
}
\concept{validators}
