% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateLocation}
\alias{validateLocation}
\title{Clears and validates the provided location}
\usage{
validateLocation(location)
}
\arguments{
\item{location}{a string representing the domain subfolder
where the appication is deployed.}
}
\value{
A validated version of the location
}
\description{
Clears and validates the provided location
}
\seealso{
[validateDomain()], [validateIcon()]

Other validators: 
\code{\link{validateDomain}()},
\code{\link{validateIcon}()}
}
\concept{validators}
