% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportPWA.R
\name{supportPWA}
\alias{supportPWA}
\title{Adds PWA app support.}
\usage{
supportPWA(
  base_url,
  title = "My Shiny PWA",
  location = "",
  icon = NULL,
  color = "#bada55",
  offline_page = NULL,
  offline_message = "Looks like you are offline :O"
)
}
\value{
A UI definition that can be passed to the [shinyUI] function.
}
\description{
Adds PWA app support.
}
\note{
See \url{https://css-tricks.com/snippets/css/complete-guide-grid/}
  for additional details on using css grids
}
\examples{
supportPWA()

}
